requirements.txt
astroid==2.3.2
autopep8==1.4.4
dj-database-url==0.5.0
Django==2.2.6
djangorestframework==3.10.3
gunicorn==19.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.1
whitenoise==4.1.4
wrapt==1.11.2
